requirement.txt
django
djangorestframework
beautifulsoup
urllib
sqlparse